Fix return type of text_to_lsn() function

+++ pool_worker_child.c
@@ -67,7 +67,7 @@ static volatile sig_atomic_t reload_config_request = 0;
 static void establish_persistent_connection(void);
 static void discard_persistent_connection(void);
 static void check_replication_time_lag(void);
-static long text_to_lsn(char *text);
+static unsigned long long int text_to_lsn(char *text);
 static RETSIGTYPE my_signal_handler(int sig);
 static RETSIGTYPE reload_config_handler(int sig);
 static void reload_config(void);
@@ -314,7 +314,7 @@ static void check_replication_time_lag(void)
 /*
  * Convert logid/recoff style text to 64bit log location (LSN)
  */
-static long text_to_lsn(char *text)
+static unsigned long long int text_to_lsn(char *text)
 {
 	unsigned int xlogid;
 	unsigned int xrecoff;
